keep track of data array length
git-svn-id: svn://svn.code.sf.net/p/udt-java/code@72 bee50969-7b09-4b63-a8b0-f2441a869d12

+++ udt-java/trunk/src/main/java/udt/packets/DataPacket.java
@@ -78,7 +78,7 @@ public class DataPacket implements UDTPacket, Comparable<UDTPacket>{
 		return this.data;
 	}
 
-	public double getLength(){
+	public int getLength(){
 		return dataLength;
 	}
